crewAIWithRag/requirements.txt
crewai==0.55.2
crewai-tools==0.12.0
poetry==1.8.3
fastapi==0.112.0
uvicorn==0.30.6
fpdf==1.7.2
pdfminer.six
nltk==3.9.1